Refactor: Remove `testOnly` attribute from debug manifest
The `android:testOnly="true"` attribute in the debug manifest has been removed. This change likely aims to allow easier installation of debug builds on non-development devices or emulators for testing purposes.

Additionally, this commit includes the removal of several intermediate build files related to AndroidTest and debug builds. These are likely byproducts of the manifest change and routine build cleanup.